\name{dfpk2-package}
\alias{dfpk2-package}
\alias{dfpk2}
\docType{package}
\title{
\packageTitle{dfpk2}
}
\description{
\packageDescription{dfpk2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dfpk2}
\packageIndices{dfpk2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dfpk2}

Maintainer: \packageMaintainer{dfpk2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\keyword{ Dose-finding }
\keyword{ Bayesian inference }

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
