% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{dosefinding-class}
\alias{dosefinding-class}
\title{An S4 class to represent a dosefinding.}
\description{
An S4 class to represent a dosefinding.
}
\section{Slots}{

\describe{
\item{\code{model}}{A character string to specify the working model used in the method.}

\item{\code{pid}}{Patient ID provided in the study.}

\item{\code{N}}{Number of patients.}

\item{\code{time}}{The time sampling.}

\item{\code{dose}}{dose levels.}

\item{\code{conc}}{Concentration of the drug.}

\item{\code{p_0}}{Skeleton of CRM.}

\item{\code{L}}{A threshold set before starting the trial.}

\item{\code{nchains}}{Number of chains.}

\item{\code{niter}}{Number of iterations.}

\item{\code{nadapt}}{Number of warmup iterations.}

\item{\code{new_dose}}{The next maximum tolerated dose.}

\item{\code{theta}}{The toxicity (probability) target.}

\item{\code{dose_levels}}{The dose levels of drug for each patient during the trial.}

\item{\code{toxicity}}{The toxicity outcome.}

\item{\code{AUCs}}{The AUCs.}
}}

