% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{methods}
\name{plot,scen,missing-method}
\alias{plot,scen,missing-method}
\title{The graphical representation of the drug's concentration in the plasma at time t after the drug administration.}
\usage{
\S4method{plot}{scen,missing}(x, y = NA, ...)
}
\arguments{
\item{x}{a "scen" object.}

\item{y}{the "y" argument is not used in the plot-method for "scen" object.}

\item{\dots}{other arguments to the \code{\link[=graphics]{plot.default}} function can be passed here.}
}
\description{
The graphical representation of the drug's concentration in the plasma at time t after the drug administration.
}
\author{
Artemis Toumazi \email{artemis.toumazi@inserm.fr}, Moreno Ursino \email{moreno.ursino@inserm.fr}, Sarah Zohar \email{sarah.zohar@inserm.fr}
}
\references{
Ursino, M., et al, (2017) Dose-finding methods for Phase I clinical trials using pharmacokinetics in small populations, Biometrical Journal.
}

