% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{methods}
\name{plot.conc,scen-method}
\alias{plot.conc,scen-method}
\title{The graphical representation of the drug's concentration in the plasma at time t after the drug administration.}
\usage{
\S4method{plot.conc}{scen}(x, ...)
}
\arguments{
\item{x}{A "scen" object.}

\item{...}{Other parameters to be passed through to plotting functions. They are ignored in this function.}
}
\description{
The graphical representation of the drug's concentration in the plasma at time t after the drug administration.
}

