% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\name{plot.conc}
\alias{plot.conc}
\title{The Generic function of the graphical representation of dosefinding.}
\usage{
plot.conc(x, ...)
}
\arguments{
\item{x}{A dosefinding object.}

\item{...}{Other parameters to be passed through to plotting functions. They are ignored in this function.}
}
\description{
The Generic function of the graphical representation of dosefinding.
}

