% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{methods}
\name{plot.dose,dosefinding-method}
\alias{plot.dose,dosefinding-method}
\title{The graphical representation of dosefinding.}
\usage{
\S4method{plot.dose}{dosefinding}(x, ...)
}
\arguments{
\item{x}{A dosefinding object.}

\item{...}{Other parameters to be passed through to plotting functions. They are ignored in this function.}
}
\description{
The graphical representation of dosefinding.
}

