% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{methods}
\name{plotConc,scen-method}
\alias{plotConc,scen-method}
\title{The graphical representation of the drug's concentration in the plasma at time t after the drug administration.}
\usage{
\S4method{plotConc}{scen}(x, ...)
}
\arguments{
\item{x}{A "scen" object.}

\item{...}{Other parameters to be passed through to plotting functions. They are ignored in this function.}
}
\description{
The graphical representation of the drug's concentration in the plasma at time t after the drug administration.
}
\author{
Artemis Toumazi \email{artemis.toumazi@inserm.fr}, Moreno Ursino \email{moreno.ursino@inserm.fr}, Sarah Zohar \email{sarah.zohar@inserm.fr}
}
\references{
Ursino, M., et al, (2016) Dose-finding methods using pharmacokinetics in small populations (under review).
}
\seealso{
\code{\link{plotDose}}
}

