% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\name{plotConc}
\alias{plotConc}
\title{The generic function of plot.conc}
\usage{
plotConc(x, ...)
}
\arguments{
\item{x}{A "scen" object.}

\item{...}{Other parameters to be passed through to plotting functions. They are ignored in this function.}
}
\description{
A graphical representation of drug's concentration in the plasma at time t.
}
\author{
Artemis Toumazi \email{artemis.toumazi@inserm.fr}, Moreno Ursino \email{moreno.ursino@inserm.fr}, Sarah Zohar \email{sarah.zohar@inserm.fr}
}
\references{
Ursino, M., et al, (2016) Dose-finding methods using pharmacokinetics in small populations (under review).
}
\seealso{
\code{\link{plotDose}}
}

