% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{methods}
\name{plotDose,dosefinding-method}
\alias{plotDose,dosefinding-method}
\title{The graphical representation of dose escalation for each patient in the trial.}
\usage{
\S4method{plotDose}{dosefinding}(x, ...)
}
\arguments{
\item{x}{A "dosefinding" object.}

\item{...}{Other parameters to be passed through to plotting functions. They are ignored in this function.}
}
\description{
The graphical representation of dose escalation for each patient in the trial.
}

