% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{scen-class}
\alias{scen-class}
\title{An S4 class to represent a simulated scenarios}
\description{
An S4 class to represent a simulated scenarios
}
\section{Slots}{

\describe{
\item{\code{param_pk}}{The subject's PK parameters.}

\item{\code{n_pk}}{The length of time sampling.}

\item{\code{time}}{The time sampling.}

\item{\code{N}}{Number of patients.}

\item{\code{doses}}{The dose levels.}

\item{\code{limit_tox}}{Threshold in the toxicity.}

\item{\code{omega2}}{A standard deviation of CL and V equals to.}

\item{\code{omega_a}}{A standard deviation omega_a.}

\item{\code{conc}}{Concentration of the drug.}

\item{\code{conc_pred}}{The predicted concentration.}

\item{\code{tox}}{The toxicity outcomes.}

\item{\code{tab}}{The data summary of scenarios' outcomes.}

\item{\code{parameters}}{The PK parameter's estimations of each patient.}

\item{\code{alfa_AUC}}{AUC with sensitivity parameter.}
}}

