\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,dosefinding-method}
\alias{show,scen-method}
\alias{show,dose-method}
\title{S4 Methods for Function \code{show}}
\description{
S4 Methods for function \code{show}.
}

\section{Methods}{
\describe{

\item{\code{signature(object = "dosefinding")}}{
S4 method to store and present the dose-finding results.
}

\item{\code{signature(object = "scen")}}{
S4 method to store and present the simulated datasets. 
}

\item{\code{signature(object = "dose")}}{
S4 method to store and present the next recommended dose level in an ongoing trial. 
}
}}
\keyword{methods}
