\name{stan_f}
\alias{stan_f}
\docType{data}

\title{
The data stan_f includes all the Stan models that dfpk package uses. 
}

\description{
	A character vector with paths to .stan files to include in the package.
}

\usage{data("stan_f")}

\references{
Ursino, M., et al, (2017) Dose-finding methods for Phase I clinical trials using pharmacokinetics in small populations, Biometrical Journal.
}

\examples{
data(stan_f)
## maybe str(stan_f) ; plot(stan_f) ...
}

\keyword{datasets}
