% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\name{plotDose}
\alias{plotDose}
\title{The generic function of the plot.dose}
\usage{
plotDose(x, ...)
}
\arguments{
\item{x}{A "dosefinding" object.}

\item{...}{Other parameters to be passed through to plotting functions. They are ignored in this function.}
}
\description{
A graphical representation of dose escalation for each patient in the trial.
}

