% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{scen-class}
\alias{scen-class}
\title{An S4 class to represent a simulated scenarios.}
\description{
An S4 class to represent a simulated scenarios.
}
\section{Slots}{

\describe{
\item{\code{PKparameters}}{The subject's PK parameters.}

\item{\code{nPK}}{The length of time sampling.}

\item{\code{time}}{The time sampling.}

\item{\code{N}}{The total number of patients.}

\item{\code{doses}}{The doses levels of the drug.}

\item{\code{preal}}{The real probability of toxicity.}

\item{\code{limitTox}}{Threshold in the toxicity.}

\item{\code{omegaIIV}}{A standard deviation of clearance (CL) and volume (V).}

\item{\code{omegaAlpha}}{A standard deviation omegaAlpha.}

\item{\code{conc}}{The concentration of the drug (concentration without error).}

\item{\code{concPred}}{The predicted concentration of the drug (concentration with error).}

\item{\code{tox}}{The toxicity outcomes.}

\item{\code{tab}}{The data summary of scenarios' outcomes.}

\item{\code{parameters}}{The PK parameter's estimations of each patient.}

\item{\code{alphaAUC}}{AUC with sensitivity parameter.}
}}

