% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{scen-class}
\alias{scen-class}
\title{An S4 class to represent a simulated scenarios.}
\description{
An S4 class to represent a simulated scenarios.
}
\section{Slots}{

\describe{
\item{\code{param_pk}}{The subject's PK parameters.}

\item{\code{n_pk}}{The length of time sampling.}

\item{\code{time}}{The time sampling.}

\item{\code{N}}{The total number of patients.}

\item{\code{doses}}{The doses levels of the drug.}

\item{\code{limit_tox}}{Threshold in the toxicity.}

\item{\code{omega2}}{A standard deviation of CL and V.}

\item{\code{omega_a}}{A standard deviation omega_a.}

\item{\code{conc}}{The concentration of the drug (concentration without error).}

\item{\code{conc_pred}}{The predicted concentration of the drug (concentration with error).}

\item{\code{tox}}{The toxicity outcomes.}

\item{\code{tab}}{The data summary of scenarios' outcomes.}

\item{\code{parameters}}{The PK parameter's estimations of each patient.}

\item{\code{alfa_AUC}}{AUC with sensitivity parameter.}
}}

