\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,dosefinding-method}
\alias{show,scen-method}
\title{S4 Methods for Function \code{show}}
\description{
S4 Methods for function \code{show}.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "dosefinding")}}{
S4 method of dosefinding.
}

\item{\code{signature(object = "scen")}}{
S4 method of simulated scenarios. 
}
}}
\keyword{methods}
