\name{stan_f}
\alias{stan_f}
\docType{data}

\title{
The data stan_f includes all the Stan models that dfpk package uses. 
}

\description{
	A character vector with paths to .stan files to include in the package.
}

\usage{data("stan_f")}

\references{
Ursino, M., et al, (2017) Dose-finding methods for Phase I clinical trials using pharmacokinetics in small populations, Biometrical Journal, <doi:10.1002/bimj.201600084>.

Toumazi, A., et al, (2018) dfpk: An R-package for Bayesian dose-finding designs using pharmacokinetics (PK) for phase I clinical trials, Computer Methods and Programs in Biomedicine, <doi:10.1016/j.cmpb.2018.01.023>.
}

\examples{
data(stan_f)
## maybe str(stan_f) ; plot(stan_f) ...
}

\keyword{datasets}
