\name{stan_f}
\alias{stan_f}
\docType{data}
\title{
The data stan_f includes all the Stan models that dfpk package uses.
}

\description{
	A character vector with paths to .stan files to include in the package.
}

\usage{data("stan_f")}

\references{
Ursino, M., et al, (2016) Dose-finding methods using pharmacokinetics in small populations (under review).
}

\examples{
data(stan_f)
## maybe str(stan_f) ; plot(stan_f) ...
}

\keyword{datasets}
