\name{stan_f}
\alias{stan_f}
\docType{data}
\title{
Stan models
}
\description{
	A character vector with paths to .stan files to include in the package.
}
\usage{data("stan_f")}
\format{
  The format is:
 chr [1:5] "/Volumes/USB DISK/Inserm_1138/Artemis' code/STAN_models/RPackage/logit_cov2.stan" ...
}
\references{
Moreno et al, 2016
}
\examples{
data(stan_f)
## maybe str(stan_f) ; plot(stan_f) ...
}
\keyword{datasets}
